# sqlbase: interleave descriptors and key coding