--- a/sql/sqlbase/encoded_datum.pb.go
+++ b/sql/sqlbase/encoded_datum.pb.go
@@ -17,6 +17,7 @@ It has these top-level messages:
 	ForeignKeyReference
 	ColumnDescriptor
 	ColumnFamilyDescriptor
+	InterleaveDescriptor
 	IndexDescriptor
 	DescriptorMutation
 	TableDescriptor