--- a/sql/sqlbase/structured.go
+++ b/sql/sqlbase/structured.go
@@ -1301,8 +1301,3 @@ func (desc *Descriptor) GetName() string {
 		return ""
 	}
 }
-
-// IndexKeyPrefix returns the prefix for the table-index reference.
-func (t ForeignKeyReference) IndexKeyPrefix() []byte {
-	return MakeIndexKeyPrefix(t.Table, t.Index)
-}