Implements the encoding described in
https://github.com/cockroachdb/cockroach/blob/master/docs/RFCS/sql_interleaved_tables.md

No change in benchmarks for non-interleaved tables.

name                       old time/op    new time/op    delta
Insert1000_Cockroach-8       9.73ms ± 4%    9.71ms ± 3%    ~             (p=0.690 n=5+5)
Update1000_Cockroach-8       28.6ms ± 4%    29.0ms ± 2%    ~             (p=0.310 n=5+5)
Delete1000_Cockroach-8       49.8ms ± 4%    48.2ms ± 5%    ~             (p=0.222 n=5+5)
Scan1000_Cockroach-8         2.28ms ± 2%    2.24ms ± 2%    ~             (p=0.222 n=5+5)
WideTable1000_Cockroach-8     109ms ± 4%     112ms ± 2%    ~             (p=0.151 n=5+5)

name                       old alloc/op   new alloc/op   delta
Insert1000_Cockroach-8       4.18MB ± 0%    4.18MB ± 0%    ~             (p=0.413 n=4+5)
Update1000_Cockroach-8       6.55MB ± 0%    6.54MB ± 0%    ~             (p=0.690 n=5+5)
Delete1000_Cockroach-8       3.19MB ± 0%    3.19MB ± 0%    ~             (p=0.690 n=5+5)
Scan1000_Cockroach-8          317kB ± 0%     317kB ± 0%  +0.04%          (p=0.008 n=5+5)
WideTable1000_Cockroach-8    21.2MB ± 0%    21.2MB ± 0%    ~             (p=0.690 n=5+5)

name                       old allocs/op  new allocs/op  delta
Insert1000_Cockroach-8        20.4k ± 0%     20.4k ± 0%    ~             (p=0.540 n=4+5)
Update1000_Cockroach-8        29.5k ± 0%     29.5k ± 0%    ~             (p=0.056 n=5+5)
Delete1000_Cockroach-8        21.6k ± 0%     21.6k ± 0%    ~             (p=0.810 n=5+5)
Scan1000_Cockroach-8          2.30k ± 0%     2.30k ± 0%    ~     (all samples are equal)
WideTable1000_Cockroach-8      223k ± 0%      223k ± 0%    ~             (p=0.690 n=5+5)

For #2972.

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/cockroachdb/cockroach/7755)
<!-- Reviewable:end -->