@RaduBerinde 2016-07-11
Review status: 0 of 20 files reviewed at latest revision, 4 unresolved discussions, all commit checks successful.

---

*[sql/sqlbase/structured.proto, line 202 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/7755#-KMR2zD70_oo9OMP_0AF:-KMR2zD70_oo9OMP_0AG:555969311) ([raw file](https://github.com/cockroachdb/cockroach/blob/376daf80a30445cae91fd9a7c08860b54433b85d/sql/sqlbase/structured.proto#L202)):*
> ```protobuf
> 
> // InterleaveDescriptor represents an index (either primary or secondary) that
> // is interleaved into another table's data.
> ```

I think the example below should be moved up here and should mention how the entire InterleaveDescriptor looks and for which table it is set. (In fact there will be one set for table 3, and another one set for table 2 right?)

---

*[sql/sqlbase/structured.proto, line 206 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/7755#-KMR398BYA7112SMCwSc:-KMR398BYA7112SMCwSd:2115764884) ([raw file](https://github.com/cockroachdb/cockroach/blob/376daf80a30445cae91fd9a7c08860b54433b85d/sql/sqlbase/structured.proto#L206)):*
> ```protobuf
>   message Component {
>     // ParentID the ID of the table being interleaved into.
>     optional uint32 parent_id = 1 [(gogoproto.nullable) = false,
> ```

`ancestor_table_id` (and `ancestor_index_id`)? can also be just `table_id`/`index_id` if the struct is called Ancestor

---

*[sql/sqlbase/structured.proto, line 231 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/7755#-KMR2TaICuLE1srF-CHI:-KMR2TaICuLE1srF-CHJ:1562179118) ([raw file](https://github.com/cockroachdb/cockroach/blob/376daf80a30445cae91fd9a7c08860b54433b85d/sql/sqlbase/structured.proto#L231)):*
> ```protobuf
>   // an encoded key. The last element is interleaved into the second to last,
>   // and so on.
>   repeated Component components = 2 [(gogoproto.nullable) = false];
> ```

Maybe it would make more sense for this to be called ancestors instead of components?

The comment should mention that they are always in the far-to-near ancestor order (e.g. grand-grand-parent, grand-parent, parent).

---

*[sql/sqlbase/structured.proto, line 249 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/7755#-KMR3_8RD6h_vSwFzDdM:-KMR3_8RD6h_vSwFzDdN:-830697968) ([raw file](https://github.com/cockroachdb/cockroach/blob/376daf80a30445cae91fd9a7c08860b54433b85d/sql/sqlbase/structured.proto#L249)):*
> ```protobuf
>   optional uint32 parent_id = 4 [(gogoproto.nullable) = false,
>       (gogoproto.customname) = "ParentID", (gogoproto.casttype) = "ID"];
>   repeated InterleaveDescriptor interleaves = 24 [(gogoproto.nullable) = false,
> ```

Doesn't this belong into IndexDescriptor instead of TableDescriptor?

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/7755)*
<!-- Sent from Reviewable.io -->