@paperstreet 2016-07-12
Thanks @raduberinde! RFAL

---

Review status: 0 of 20 files reviewed at latest revision, 4 unresolved discussions, some commit checks pending.

---

*[sql/sqlbase/structured.proto, line 202 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/7755#-KMR2zD70_oo9OMP_0AF:-KMUdc94x-t-NvV-lfZR:201294969) ([raw file](https://github.com/cockroachdb/cockroach/blob/376daf80a30445cae91fd9a7c08860b54433b85d/sql/sqlbase/structured.proto#L202)):*
<details><summary><i>Previously, RaduBerinde wrote…</i></summary>
> I think the example below should be moved up here and should mention how the entire InterleaveDescriptor looks and for which table it is set. (In fact there will be one set for table 3, and another one set for table 2 right?)
</details>
Fair point

---

*[sql/sqlbase/structured.proto, line 206 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/7755#-KMR398BYA7112SMCwSc:-KMUdUoWufOoQa3VCqLP:-499146009) ([raw file](https://github.com/cockroachdb/cockroach/blob/376daf80a30445cae91fd9a7c08860b54433b85d/sql/sqlbase/structured.proto#L206)):*
<details><summary><i>Previously, RaduBerinde wrote…</i></summary>
> `ancestor_table_id` (and `ancestor_index_id`)? can also be just `table_id`/`index_id` if the struct is called Ancestor
</details>
Done.

---

*[sql/sqlbase/structured.proto, line 231 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/7755#-KMR2TaICuLE1srF-CHI:-KMUdYFtrOpH-0dKqfBR:1337007447) ([raw file](https://github.com/cockroachdb/cockroach/blob/376daf80a30445cae91fd9a7c08860b54433b85d/sql/sqlbase/structured.proto#L231)):*
<details><summary><i>Previously, RaduBerinde wrote…</i></summary>
> Maybe it would make more sense for this to be called ancestors instead of components?
> 
> The comment should mention that they are always in the far-to-near ancestor order (e.g. grand-grand-parent, grand-parent, parent).
</details>
Sure, I like Ancestor

---

*[sql/sqlbase/structured.proto, line 249 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/7755#-KMR3_8RD6h_vSwFzDdM:-KMUe2JYF1xOPCt52wUR:-1571609979) ([raw file](https://github.com/cockroachdb/cockroach/blob/376daf80a30445cae91fd9a7c08860b54433b85d/sql/sqlbase/structured.proto#L249)):*
<details><summary><i>Previously, RaduBerinde wrote…</i></summary>
> Doesn't this belong into IndexDescriptor instead of TableDescriptor?
</details>
Ah, that does model the at-most-one interleave per index more correctly. I did it the other way because I was thinking the code would be a little cleaner, but that's likely the wrong tradeoff. Fixed.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/7755#-:-KMVMLIMmWZz4D-j3m7A:-1529773659)*
<!-- Sent from Reviewable.io -->